+++ cafe-orders/src/components/waiters/WaitersForm.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useState } from 'react';
 import { withRouter, useHistory } from 'react-router-dom';
 import { Formik, Form, Field } from 'formik';
 import * as yup from 'yup';
@@ -9,6 +9,10 @@ import SaveIcon from '@material-ui/icons/Save';
 import CancelIcon from '@material-ui/icons/Cancel';
 import DeleteIcon from '@material-ui/icons/Delete';
 import TextField from '@material-ui/core/TextField';
+import Dialog from '@material-ui/core/Dialog';
+import DialogActions from '@material-ui/core/DialogActions';
+import DialogContent from '@material-ui/core/DialogContent';
+import DialogTitle from '@material-ui/core/DialogTitle';
 import { connect } from 'react-redux';
 import { saveWaiter, deleteWaiter } from '../../store/actions/waiters';
 
@@ -19,21 +23,41 @@ const BLANK = {
     surname: '',
     notation: '',
 }
-function WaitersForm({ match: { params: { id } }, waiters, saveWaiter, deleteWaiter }) {
+function WaitersForm(
+    { match: { params: { id } },
+        waiters,
+        orders,
+        saveWaiter,
+        deleteWaiter
+    }) {
 
     const history = useHistory();
+    const [modalVisible, setModalVisible] = useState(false);
 
+    const closeModal = () => setModalVisible(false);
     const selectedItem = (id) => waiters.find(item => item._id === id);
     const getItemForForm = (id) => id === 'new' ? BLANK : selectedItem(id);
     const onCancel = () => history.goBack();
-    const saveButtonStatus = (isValid, dirty) =>  !(isValid && dirty);
+    const saveButtonStatus = (isValid, dirty) => !(isValid && dirty);
     const onFormSubmit = (data) => {
         saveWaiter(data);
         history.goBack();
     }
+    const findForDelete = () => {
+        let array = orders.filter((item => item.waiterID === id));
+        let result = [];
+        array.forEach((item, index) => {
+            result.push(`${index+1}. Order from ${item.date} in the amount of ${item.sum}`)
+        });
+        return { length: array.length, result }
+    }
     const onDelete = () => {
-        deleteWaiter(id);
-        history.goBack();
+        if (findForDelete().length !== 0) {
+            setModalVisible(true);
+        } else {
+            deleteWaiter(id);
+            history.goBack();
+        }
     }
     const waiterSchema = yup.object().shape({
         name: yup.string()
@@ -142,12 +166,38 @@ function WaitersForm({ match: { params: { id } }, waiters, saveWaiter, deleteWai
                     </Grid>
                 </Form>
             </Formik>
+            {
+                modalVisible
+                    ? <>
+                        <Dialog
+                            open={modalVisible}
+                            onClose={closeModal}
+                            aria-labelledby="alert-dialog-title"
+                            aria-describedby="alert-dialog-description"
+                        >
+                            <DialogTitle id="alert-dialog-title">{"Deletion aborted."}</DialogTitle>
+                            <DialogContent>
+                                <div>
+                                Found related data. First, delete the records containing this item.
+                                {findForDelete().result.map((item, index) => <p key={index}>{item}</p>)}
+                                </div>
+                            </DialogContent>
+                            <DialogActions>
+                                <Button onClick={closeModal} color="primary" autoFocus>
+                                    Ok
+                                </Button>
+                            </DialogActions>
+                        </Dialog>
+                    </>
+                    : null
+            }
         </>
     )
 }
 
-const mapStateToProps = ({ waiters: { items } }) => ({
-    waiters: items,
+const mapStateToProps = ({ waiters, orders }) => ({
+    waiters: waiters.items,
+    orders: orders.items,
 });
 const mapDispatchToProps = {
     saveWaiter,
@@ -164,5 +214,22 @@ const styles = {
     },
     error: {
         color: 'red'
+    },
+    modalBackground: {
+        background: 'grey',
+        opacity: '0.3',
+        width: '100%',
+        height: '100%',
+        position: 'absolute',
+        top: '0',
+        left: '0',
+    },
+    modal: {
+        position: 'absolute',
+        width: '100%',
+        height: '100%',
+        top: '0',
+        left: '0',
+        zIndex: '2',
     },
 }
